Fix message/mailfolder examples and add multiple child folders method
