+++ src/child_folders/mod.rs
@@ -1,3 +1,4 @@
+mod manual_request;
 mod request;
-
+pub use manual_request::*;
 pub use request::*;
